SW-3765: greybus_protocols.h Add SVC_TIMESYNC_PING
In order to verify timesync functionality we require a timesync-ping
operation where the AP can command the SVC to initiate a single strobe of
downstream timesync slaves, and report the frame-time at the strobe. Ping
will only be valid after the system has transitioned to a TIMESYNC_ACTIVE
state.

In the active state each timesync slave will graph a single incoming SVC
strobe as a ping and will store its frame time. The AP will then gather
each last-event frame-time for presentation to user-space in the AP and/or
further automation based on the reported frame-times at the SVC ping.

This patch adds the SVC ping command definition to greybus_protocols.h.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Vaibhav Hiremath <vaibhav.hiremath@linaro.org>