SW-3744: greybus_protocols.h Add SVC_TIMESYNC_WD_PINS_ACQUIRE/RELEASE
Its necessary to establish an initial state on the wake-detect lines before
enabling timesync on APB/GPB. We can't rely on the following sequence

- SVC timesync-enable (implicit GPIO output state)
- APB/GPB timesync-enable

Since we can't guarantee all APBs/GPBs will receive the timesync start
command prior to the SVC sending its strobe set. As a result we need to.

- Establish the initial state of the WDn lines as outputs from SVC
- Enable timesync RX on APB/GPB
- Enable timesync on SVC - which enables the output strobe-train from SVC
  SVC will release the WDn lines automatically once the synchronization
  operation completes.
- Gather authoritative time at each SVC output pulse
- Diseminate authoritative time to each interested party.

This patch adds the relevant SVC command to greybus_protocols.h

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Vaibhav Hiremath <vaibhav.hiremath@linaro.org>