SW-3335: Add gb_timesync_frame_time_to_timespec()
This patch adds gb_timesync_frame_time_to_timespec() a routine that will
convert from a given frame-time to a ktime/timespec within an envelope of
about 16 seconds. The purpose of this routine is to enable reporting of a
frame-time from a Module such as a Camera Module and to allow the AP to
then convert this timestamp into a Linux-native timestamp such as ktime.
This is useful and required in the v4l layer.

At 19.2MHz the accuracy of this conversion is about .3 femtoseconds per
count, which means at a 1 second offset from the reference the cumlative
error is about 1.59 nanoseconds. 1.59 nanoseconds is still less than 1
clock's worth of error @ 19.2MHz where each clock is 52.0833~ nanoseconds.

We're aiming for a maximum error rate of 30 nanoseconds which means at the
clock rate we are running at the conversion from a frame-time to a Linux
ktime/timespec can be plus-or-minus about 17 seconds from the reference
frame-time/ktime pair before the routine will refuse to convert.

A realistic use-case for this routine is envisaged to be

- Greybus message received
- Some processing takes place taking milliseconds
- Call into this routine is made
- Actual time between even in Module and conversion in AP < 1 second
- Error rate in conversion at 1.59 nanoseconds is less than 1 clock
  @ 19.2MHz

This rountine is not designed to allow for conversions for events with
large gaps i.e. over the 10-17 second range.

Adds:
- timesync.c::gb_timesync_frame_time_to_timespec(u64 frame_time,
                                                 struct timespec *ts)
  - frame-time is a system frame-time.
  - ts is an output parameter which represents the converted frame-time
    as a CLOCK_MONOTONIC timespec value.
  - Returns 0 on success or a negative number indicating the type of
    error on failure.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>